feat: introduce provider
add the Provider with search and download function.
add a FakeProvider implementation for search and download.
add a EodagProvider implementation for download.

EodagProvider.search isn't implemented yet.
There is still work to do on EodagProvider.download :
* the authentication isn't done currently
* verify the compliance with the CADIP ICD

The provider isn't used on the CADIP service endpoint currently. It will be branched later.

Refs: #RSPY-16