fix: data_retriever tests
data_retriever unit tests was failing because code has been commented for integration with the api layer.

The api layer only use the provider layer so I updated so that it doesn't use the data retriever anymore but use directly the EodagProvider.

Once done, I was able to restore the data_retriever back to its normal state.