feat: add the user catalog middleware
The rs-server-common dependency isn't working anymore.
It needs a pydantic v2.*
but stac-fastapi needs a pydantic v1.*

the dependency to rs-server-common has jut been removed since it wasn't used at all.

Some work is in progress on stac-fastapi
to upgrade to pydantic v2+
We could follow this subject
and update stac-fastapi-pgstac eventually
once the subject has been handled
if we need it.

Refs: #RSPY-78